feat: add an `aws-sdk` feature enabling aws-sdk integration
This is analogous to the `rusoto` feature, with the same
implementations:

- aws-sdk's `ProvideCredentials` for `SsoFlowBuilder` and `SsoFlow`.
- `From<SessionCredentials>` for aws-sdk's `Credentials`.

`ProvideCredentials` requires implementors to also implement `Debug`,
which required some crate types to implement `Debug`. Sadly this can't
be derived for `SsoFlowBuilder` or `SsoFlow` since we want closures to
be usable as verification prompts, and closures do not implement
`Debug`.